app/requirements.txt
aiodynamo==24.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
authx==1.4.1
billiard==4.2.1
boto3==1.35.91
botocore==1.35.91
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jmespath==1.0.1
kombu==5.4.2
mailersend==0.5.8
multidict==6.1.0
passlib==1.7.4
pika==1.3.2
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
setuptools==69.5.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
yarl==1.18.3
